All tests passing, some new tests still need to be added

+++ src/source/PeerConnection/JitterBuffer.h
@@ -32,7 +32,9 @@ typedef struct {
     UINT16 tailSequenceNumber;
     UINT32 headTimestamp;
     UINT32 tailTimestamp;
-    UINT32 maxLatency;
+    //this is set to U64 even though rtp timestamps are U32
+    //in order to allow calculations to not cause overflow
+    UINT64 maxLatency;
     UINT64 customData;
     UINT32 clockRate;
     BOOL started;
